requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
cssselect==1.2.0
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
jmespath==1.0.1
lxml==4.9.3
magic-filter==1.0.11
multidict==6.0.4
packaging==23.1
parsel==1.8.1
python-decouple==3.8
pytz==2023.3
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.4
w3lib==2.1.1
yarl==1.9.2
